# Remove Unsafe usage from `GifDecoderCore` and optimize loops